% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beer_Song.R
\name{Beer_Song}
\alias{Beer_Song}
\title{Write a Beautiful Song}
\usage{
Beer_Song(n, liquid, vessel, container)
}
\arguments{
\item{n}{a number}

\item{liquid}{a type of beer}

\item{vessel}{a type of vessel}

\item{container}{a type of container}
}
\value{
Beer_Song
}
\description{
Using this function can write a song of beautiful lyrics.
You can change the factors to get different lyrics.
}
\examples{
Beer_Song(n = 10, liquid = "beer", vessel = "bottles", container = "shelf")
}
