% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Titanic.R
\docType{data}
\name{Titanic}
\alias{Titanic}
\title{Survival of passengers on the Titanic.}
\format{A 4-dimensional array results from cross-tabulating 2201 observations on 4 variables:
\describe{
  \item{Name}{Name, devided into "Class", "Sex", "Age", "Survived"}
  \item{Level}{Level, the answers respond to Name}
}}
\source{
\url{https://www.encyclopedia-titanica.org/}
}
\usage{
Titanic
}
\description{
This data set provides information on the fate of passengers
on the fatal maiden voyage of the ocean liner ‘Titanic’,
summarized according to economic status(class), sex, age and survival.
}
\keyword{datasets}
